+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Fixed
+
+- `usePeerState()` not being able to store more than the latest peer connection
+
 ## [0.1.0][] - 2019-01-28
 
 - initial release
 
 <!-- prettier-ignore -->
-
-
 [Unreleased]: https://github.com/madou/react-peer/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/madou/react-peer/tree/v0.1.0